Reuse default toolbox properties for workflow tool box component
