+++ client/galaxy/scripts/components/Panels/ToolBox.vue
@@ -95,7 +95,6 @@ export default {
     },
     created() {
         this.toolsLayout = getToolsLayout();
-        console.log(this.toolsLayout);
     },
     methods: {
         setResults(results) {
